merge before tag

See merge request niryo/niryo-one-s/ned_ros_stack!188

+++ niryo_robot_hardware_stack/ttl_driver/include/ttl_driver/abstract_dxl_driver.hpp
@@ -51,11 +51,11 @@ class AbstractDxlDriver : public AbstractMotorDriver
         virtual int readLoad(uint8_t id, uint16_t& present_load ) = 0;
         virtual int syncReadLoad(const std::vector<uint8_t>& id_list, std::vector<uint16_t>& load_list ) = 0;
 
-        virtual int readPID(uint8_t id, std::vector<uint32_t>& data) = 0;
+        virtual int readPID(uint8_t id, std::vector<uint16_t>& data) = 0;
         virtual int readControlMode(uint8_t id, uint8_t& control_mode) = 0;
 
         // ram write
-        virtual int writePID(uint8_t id, const std::vector<uint32_t>& data) = 0;
+        virtual int writePID(uint8_t id, const std::vector<uint16_t>& data) = 0;
         virtual int writeControlMode(uint8_t id, uint8_t data) = 0;
 
         virtual int writeLed(uint8_t id, uint8_t led_value ) = 0;
